np.int -> int (#108)

+++ pyem/vop/binary.py
@@ -39,7 +39,7 @@ def binary_volume_opening(vol, minvol):
         return vol.copy()
     lb_vol, num_objs = label(vol)
     lbs = np.arange(1, num_objs + 1)
-    v = labeled_comprehension(lb_vol > 0, lb_vol, lbs, np.sum, np.int, 0)
+    v = labeled_comprehension(lb_vol > 0, lb_vol, lbs, np.sum, int, 0)
     if minvol < 0:
         ix = np.isin(lb_vol, lbs[np.argsort(v)[minvol:]])
     else:
